requirements.txt
allure-pytest==2.8.21
allure-python-commons==2.8.21
apipkg==1.5
appdirs==1.4.4
Appium-Python-Client==0.51
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
common==0.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
entrypoints==0.3
et-xmlfile==1.0.1
execnet==1.7.1
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.4
flask-blueprint==1.3.0
Flask-Script==2.0.6
greenlet==1.1.2
idna==2.9
importlib-metadata==3.3.0
iniconfig==1.1.1
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.3
JPype1==0.7.2
jsonpath==0.82
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.3
mistune==0.8.4
MouseInfo==0.1.3
mysql==0.0.2
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.19.1
opencv-python==4.4.0.42
openpyxl==3.0.3
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
py==1.9.0
PyAutoGUI==0.9.50
pygal==2.4.0
PyGetWindow==0.0.8
Pygments==2.6.1
pymongo==3.12.1
PyMsgBox==1.0.8
PyMySQL==0.9.3
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
pyrsistent==0.16.0
PyScreeze==0.1.26
pytest==6.1.2
pytest-forked==1.3.0
pytest-rerunfailures==9.1.1
pytest-xdist==2.1.0
python-dateutil==2.8.1
PyTweening==1.0.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.3
QtPy==1.9.0
reader==1.12
requests==2.23.0
selenium==3.141.0
Send2Trash==1.5.0
sgmllib3k==1.0.0
six==1.15.0
SQLAlchemy==1.4.27
suds-jurko==0.6
terminado==0.8.3
testpath==0.4.4
toml==0.10.2
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.8
virtualenv==20.2.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==3.4.0
